Are you up for the BIZZBY iOS challenge?
Nice! You took the time to read more... This is already an extra point, and you shall be rewarded!

Here at BIZZBY we are looking for developers that want to do things properly, tidy, tested and with love.

* Show us a tidy repo with meaningful commit messages
* Write tests for the mapping/parsing logic
* Want to impress? Put some Swift!

When you come in to chat with us mention the fact that you read this! 👍